server/requirements.txt
fastapi
vosk
uvicorn
python-dotenv
pydantic
httpx
pydub
python-multipart